update get and set, add new columns on tableView

+++ src/main/resources/sql/create.sql
@@ -3,7 +3,7 @@ USE gestcursos;
 CREATE TABLE course (
                             Id int NOT NULL AUTO_INCREMENT,
                             Name varchar(60) DEFAULT NULL,
-                            Semester float DEFAULT NULL,
+                            Semester int DEFAULT NULL,
                             PRIMARY KEY (Id)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
